Add helper function for handing include URL param, use that for Connection
