+++ src/api/connections.js
@@ -2,8 +2,9 @@
 'use strict';
 
 // ---------------------- Dependencies ---->>
-
 const _ = require('lodash');
+const Helpers = require('./helpers.js');
+
 var Services = require('../services/index.js');
 var ConnectionService = Services.ConnectionService;
 
@@ -47,25 +48,13 @@ exports.del = function(req, res) {
 
 
 exports.find = function(req, res) {
-  var Model, Models;
-  var params = {};
-
-  Model = require('../models/index.js');
-  Models = new Model().models;
-
+  const Model = require('../models/index.js');
+  const models = new Model().models;
+  const params = {};
   params.where = {id: req.params.id};
-
   // ie: ?include=category&include=file&include=image
-  if (req.query.include) {
-    for (var i = 0; i < req.query.include.length; i++) {
-      var capitalize = req.query.include[i][0].toUpperCase() + req.query.include[i].slice(1);
-      includes.push(Models[capitalize]);
-    }
-    params.include = includes;
-  }
-
+  Helpers.applyModelIncludes(params, req, models);
   var Connection = new ConnectionService(params);
-
   Connection
     .find()
     .then(function(result) {
@@ -79,39 +68,11 @@ exports.find = function(req, res) {
 
 exports.findAll = function(req, res) {
   var Model = require('../models/index.js');
-  var Models = new Model().models;
-  var includes = [];
+  const models = new Model().models;
   var params = {};
-
   // ie: ?include=category&include=file&include=image
-  console.log("Req.query.include: ", req.query.include);
-  if (req.query.include) {
-    var include = req.query.include;
-    if (_.isArray(include)) {
-      includes = _.clone(req.query.include);
-    } else {
-      includes.push(include);
-    }
-    includes = _.map(includes, (include) => {
-      return {
-        model: Models[_.capitalize(include)]
-      };
-    });
-    params.include = includes;
-  }
-  /*
-  if (req.query.include) {
-    for (var i = 0; i < req.query.include.length; i++) {
-      var capitalize = req.query.include[i][0].toUpperCase() + req.query.include[i].slice(1);
-      includes.push(Models[capitalize]);
-    }
-    params.include = includes;
-  }
-  */
-  console.log("Params: ", params);
-
+  Helpers.applyModelIncludes(params, req, models);
   var Connection = new ConnectionService(params);
-
   Connection
     .findAll()
     .then(function(result) {
